se modifico el tipo de dato que retorna la funcion powerFunction cuando el exponente es negativo, asignandole float como tipo.

+++ tp1_2.c
@@ -38,7 +38,7 @@ float powerFunction(float base, int exponent)
         }
         else
         {
-            return 1 / result;
+            return (float)(1 / result);
         }
     }
 }
